Split elements and beams into separate files (#158)
* Split elements and beams into separate files along with original imports

* Fix imports in element and beam files

* Update CHANGELOG

* Setup `__init__.py` files

* Fix failing tests

* Test to see if recursive fixes autodoc

* Test `show-inheritance` to fix autodoc

* Test `autosummary` on `accelerator` subpackage

* Return to documenting (currently inexistant) `accelerator` module

* Test documenting a single element module

* Quick-and-dirty hack entering all of the element and beam modules into the current documentation by hand

* Update comments and force rebuild
