+++ .vscode/ltex.dictionary.en-GB.txt
@@ -15,3 +15,4 @@ docstring
 pytest
 isort
 docstrings
+@jp-ga
