+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - Update versions of some steps used by GitHub actions to handle Node.js 16 end-of-life (@jank324)
 - Update versions in pre-commit config (see #148) (@jank324)
+- Split `accelerator` and `beam` into separate submodules (see #158) (@jank324)
 
 ## [v0.6.3](https://github.com/desy-ml/cheetah/releases/tag/v0.6.3) (2024-03-28)
