+++ cheetah/__init__.py
@@ -1,4 +1,19 @@
-# flake8: noqa
-import cheetah.converters
-from cheetah.accelerator import *
-from cheetah.particles import *
+import cheetah.converters  # noqa: F401
+from cheetah.accelerator import (  # noqa: F401
+    BPM,
+    Aperture,
+    Cavity,
+    CustomTransferMap,
+    Dipole,
+    Drift,
+    HorizontalCorrector,
+    Marker,
+    Quadrupole,
+    RBend,
+    Screen,
+    Segment,
+    Solenoid,
+    Undulator,
+    VerticalCorrector,
+)
+from cheetah.particles import ParameterBeam, ParticleBeam  # noqa: F401
