+++ cheetah/accelerator/segment.py
@@ -1,0 +1,542 @@
+from copy import deepcopy
+from pathlib import Path
+from typing import Any, Optional, Union
+
+import matplotlib
+import matplotlib.pyplot as plt
+import torch
+from scipy import constants
+from scipy.constants import physical_constants
+from torch import Size, nn
+
+from cheetah.converters.dontbmad import convert_bmad_lattice
+from cheetah.converters.nxtables import read_nx_tables
+from cheetah.latticejson import load_cheetah_model, save_cheetah_model
+from cheetah.particles import Beam, ParticleBeam
+from cheetah.utils import UniqueNameGenerator
+
+from .custom_transfer_map import CustomTransferMap
+from .drift import Drift
+from .element import Element
+from .marker import Marker
+
+generate_unique_name = UniqueNameGenerator(prefix="unnamed_element")
+
+rest_energy = torch.tensor(
+    constants.electron_mass
+    * constants.speed_of_light**2
+    / constants.elementary_charge  # electron mass
+)
+electron_mass_eV = torch.tensor(
+    physical_constants["electron mass energy equivalent in MeV"][0] * 1e6
+)
+
+
+class Segment(Element):
+    """
+    Segment of a particle accelerator consisting of several elements.
+
+    :param cell: List of Cheetah elements that describe an accelerator (section).
+    :param name: Unique identifier of the element.
+    """
+
+    def __init__(self, elements: list[Element], name: Optional[str] = None) -> None:
+        super().__init__(name=name)
+
+        self.elements = nn.ModuleList(elements)
+
+        for element in self.elements:
+            # Make elements accessible via .name attribute. If multiple elements have
+            # the same name, they are accessible via a list.
+            if element.name in self.__dict__:
+                if isinstance(self.__dict__[element.name], list):
+                    self.__dict__[element.name].append(element)
+                else:  # Is instance of cheetah.Element
+                    self.__dict__[element.name] = [self.__dict__[element.name], element]
+            else:
+                self.__dict__[element.name] = element
+
+    def subcell(self, start: str, end: str) -> "Segment":
+        """Extract a subcell `[start, end]` from an this segment."""
+        subcell = []
+        is_in_subcell = False
+        for element in self.elements:
+            if element.name == start:
+                is_in_subcell = True
+            if is_in_subcell:
+                subcell.append(element)
+            if element.name == end:
+                break
+
+        return self.__class__(subcell)
+
+    def flattened(self) -> "Segment":
+        """
+        Return a flattened version of the segment, i.e. one where all subsegments are
+        resolved and their elements entered into a top-level segment.
+        """
+        flattened_elements = []
+        for element in self.elements:
+            if isinstance(element, Segment):
+                flattened_elements += element.flattened().elements
+            else:
+                flattened_elements.append(element)
+
+        return Segment(elements=flattened_elements, name=self.name)
+
+    def transfer_maps_merged(
+        self, incoming_beam: Beam, except_for: Optional[list[str]] = None
+    ) -> "Segment":
+        """
+        Return a segment where the transfer maps of skipable elements are merged into
+        elements of type `CustomTransferMap`. This can be used to speed up tracking
+        through the segment.
+
+        :param incoming_beam: Beam that is incoming to the segment. NOTE: This beam is
+            needed to determine the energy of the beam when entering each element, as
+            the transfer maps of merged elements might depend on the beam energy.
+        :param except_for: List of names of elements that should not be merged despite
+            being skippable. Usually these are the elements that are changed from one
+            tracking to another.
+        :return: Segment with merged transfer maps.
+        """
+        if except_for is None:
+            except_for = []
+
+        merged_elements = []  # Elements for new merged segment
+        skippable_elements = []  # Keep track of elements that are not yet merged
+        tracked_beam = incoming_beam
+        for element in self.elements:
+            if element.is_skippable and element.name not in except_for:
+                skippable_elements.append(element)
+            else:
+                if len(skippable_elements) == 1:
+                    merged_elements.append(skippable_elements[0])
+                    tracked_beam = skippable_elements[0].track(tracked_beam)
+                elif len(skippable_elements) > 1:  # i.e. we need to merge some elements
+                    merged_elements.append(
+                        CustomTransferMap.from_merging_elements(
+                            skippable_elements, incoming_beam=tracked_beam
+                        )
+                    )
+                    tracked_beam = merged_elements[-1].track(tracked_beam)
+                skippable_elements = []
+
+                merged_elements.append(element)
+                tracked_beam = element.track(tracked_beam)
+
+        if len(skippable_elements) > 0:
+            merged_elements.append(
+                CustomTransferMap.from_merging_elements(
+                    skippable_elements, incoming_beam=tracked_beam
+                )
+            )
+
+        return Segment(elements=merged_elements, name=self.name)
+
+    def without_inactive_markers(
+        self, except_for: Optional[list[str]] = None
+    ) -> "Segment":
+        """
+        Return a segment where all inactive markers are removed. This can be used to
+        speed up tracking through the segment.
+
+        NOTE: `is_active` has not yet been implemented for Markers. Therefore, this
+        function currently removes all markers.
+
+        :param except_for: List of names of elements that should not be removed despite
+            being inactive.
+        :return: Segment without inactive markers.
+        """
+        # TODO: Add check for is_active once that has been implemented for Markers
+        if except_for is None:
+            except_for = []
+
+        return Segment(
+            elements=[
+                element
+                for element in self.elements
+                if not isinstance(element, Marker) or element.name in except_for
+            ],
+            name=self.name,
+        )
+
+    def without_inactive_zero_length_elements(
+        self, except_for: Optional[list[str]] = None
+    ) -> "Segment":
+        """
+        Return a segment where all inactive zero length elements are removed. This can
+        be used to speed up tracking through the segment.
+
+        NOTE: If `is_active` is not implemented for an element, it is assumed to be
+        inactive and will be removed.
+
+        :param except_for: List of names of elements that should not be removed despite
+            being inactive and having a zero length.
+        :return: Segment without inactive zero length elements.
+        """
+        if except_for is None:
+            except_for = []
+
+        return Segment(
+            elements=[
+                element
+                for element in self.elements
+                if element.length > 0.0
+                or (hasattr(element, "is_active") and element.is_active)
+                or element.name in except_for
+            ],
+            name=self.name,
+        )
+
+    def inactive_elements_as_drifts(
+        self, except_for: Optional[list[str]] = None
+    ) -> "Segment":
+        """
+        Return a segment where all inactive elements (that have a length) are replaced
+        by drifts. This can be used to speed up tracking through the segment and is a
+        valid thing to as inactive elements should basically be no different from drift
+        sections.
+
+        :param except_for: List of names of elements that should not be replaced by
+            drifts despite being inactive. Usually these are the elements that are
+            currently inactive but will be activated later.
+        :return: Segment with inactive elements replaced by drifts.
+        """
+        if except_for is None:
+            except_for = []
+
+        return Segment(
+            elements=[
+                (
+                    element
+                    if (hasattr(element, "is_active") and element.is_active)
+                    or element.length == 0.0
+                    or element.name in except_for
+                    else Drift(element.length)
+                )
+                for element in self.elements
+            ],
+            name=self.name,
+        )
+
+    @classmethod
+    def from_lattice_json(cls, filepath: str) -> "Segment":
+        """
+        Load a Cheetah model from a JSON file.
+
+        :param filename: Name/path of the file to load the lattice from.
+        :return: Loaded Cheetah `Segment`.
+        """
+        return load_cheetah_model(filepath)
+
+    def to_lattice_json(
+        self,
+        filepath: str,
+        title: Optional[str] = None,
+        info: str = "This is a placeholder lattice description",
+    ) -> None:
+        """
+        Save a Cheetah model to a JSON file.
+
+        :param filename: Name/path of the file to save the lattice to.
+        :param title: Title of the lattice. If not provided, defaults to the name of the
+            `Segment` object. If that also does not have a name, defaults to "Unnamed
+            Lattice".
+        :param info: Information about the lattice. Defaults to "This is a placeholder
+            lattice description".
+        """
+        save_cheetah_model(self, filepath, title, info)
+
+    @classmethod
+    def from_ocelot(
+        cls,
+        cell,
+        name: Optional[str] = None,
+        warnings: bool = True,
+        device=None,
+        dtype=torch.float32,
+        **kwargs,
+    ) -> "Segment":
+        """
+        Translate an Ocelot cell to a Cheetah `Segment`.
+
+        NOTE Objects not supported by Cheetah are translated to drift sections. Screen
+        objects are created only from `ocelot.Monitor` objects when the string "BSC" is
+        contained in their `id` attribute. Their screen properties are always set to
+        default values and most likely need adjusting afterwards. BPM objects are only
+        created from `ocelot.Monitor` objects when their id has a substring "BPM".
+
+        :param cell: Ocelot cell, i.e. a list of Ocelot elements to be converted.
+        :param name: Unique identifier for the entire segment.
+        :param warnings: Whether to print warnings when objects are not supported by
+            Cheetah or converted with potentially unexpected behavior.
+        :return: Cheetah segment closely resembling the Ocelot cell.
+        """
+        from cheetah.converters.nocelot import ocelot2cheetah
+
+        converted = [
+            ocelot2cheetah(element, warnings=warnings, device=device, dtype=dtype)
+            for element in cell
+        ]
+        return cls(converted, name=name, **kwargs)
+
+    @classmethod
+    def from_bmad(
+        cls, bmad_lattice_file_path: str, environment_variables: Optional[dict] = None
+    ) -> "Segment":
+        """
+        Read a Cheetah segment from a Bmad lattice file.
+
+        NOTE: This function was designed at the example of the LCLS lattice. While this
+        lattice is extensive, this function might not properly convert all features of
+        a Bmad lattice. If you find that this function does not work for your lattice,
+        please open an issue on GitHub.
+
+        :param bmad_lattice_file_path: Path to the Bmad lattice file.
+        :param environment_variables: Dictionary of environment variables to use when
+            parsing the lattice file.
+        :return: Cheetah `Segment` representing the Bmad lattice.
+        """
+        bmad_lattice_file_path = Path(bmad_lattice_file_path)
+        return convert_bmad_lattice(bmad_lattice_file_path, environment_variables)
+
+    @classmethod
+    def from_nx_tables(cls, filepath: Union[Path, str]) -> "Element":
+        """
+        Read an NX Tables CSV-like file generated for the ARES lattice into a Cheetah
+        `Segment`.
+
+        NOTE: This format is specific to the ARES accelerator at DESY.
+
+        :param filepath: Path to the NX Tables file.
+        :return: Converted Cheetah `Segment`.
+        """
+        if isinstance(filepath, str):
+            filepath = Path(filepath)
+
+        return read_nx_tables(filepath)
+
+    @property
+    def is_skippable(self) -> bool:
+        return all(element.is_skippable for element in self.elements)
+
+    @property
+    def length(self) -> torch.Tensor:
+        lengths = torch.stack(
+            [element.length for element in self.elements],
+            dim=1,
+        )
+        return torch.sum(lengths, dim=1)
+
+    def transfer_map(self, energy: torch.Tensor) -> torch.Tensor:
+        if self.is_skippable:
+            tm = torch.eye(7, device=energy.device, dtype=energy.dtype).repeat(
+                (*self.length.shape, 1, 1)
+            )
+            for element in self.elements:
+                tm = torch.matmul(element.transfer_map(energy), tm)
+            return tm
+        else:
+            return None
+
+    def track(self, incoming: Beam) -> Beam:
+        if self.is_skippable:
+            return super().track(incoming)
+        else:
+            todos = []
+            for element in self.elements:
+                if not element.is_skippable:
+                    todos.append(element)
+                elif not todos or not todos[-1].is_skippable:
+                    todos.append(Segment([element], name="temporary_todo"))
+                else:
+                    todos[-1].elements.append(element)
+
+            for todo in todos:
+                incoming = todo.track(incoming)
+
+            return incoming
+
+    def broadcast(self, shape: Size) -> Element:
+        return self.__class__(
+            elements=[element.broadcast(shape) for element in self.elements],
+            name=self.name,
+        )
+
+    def split(self, resolution: torch.Tensor) -> list[Element]:
+        return [
+            split_element
+            for element in self.elements
+            for split_element in element.split(resolution)
+        ]
+
+    def plot(self, ax: plt.Axes, s: float) -> None:
+        element_lengths = [element.length[0] for element in self.elements]
+        element_ss = [0] + [
+            sum(element_lengths[: i + 1]) for i, _ in enumerate(element_lengths)
+        ]
+        element_ss = [s + element_s for element_s in element_ss]
+
+        ax.plot([0, element_ss[-1]], [0, 0], "--", color="black")
+
+        for element, s in zip(self.elements, element_ss[:-1]):
+            element.plot(ax, s)
+
+        ax.set_ylim(-1, 1)
+        ax.set_xlabel("s (m)")
+        ax.set_yticks([])
+
+    def plot_reference_particle_traces(
+        self,
+        axx: plt.Axes,
+        axy: plt.Axes,
+        beam: Optional[Beam] = None,
+        num_particles: int = 10,
+        resolution: float = 0.01,
+    ) -> None:
+        """
+        Plot `n` reference particles along the segment view in x- and y-direction.
+
+        :param axx: Axes to plot the particle traces into viewed in x-direction.
+        :param axy: Axes to plot the particle traces into viewed in y-direction.
+        :param beam: Entering beam from which the reference particles are sampled.
+        :param num_particles: Number of reference particles to plot. Must not be larger
+            than number of particles passed in `beam`.
+        :param resolution: Minimum resolution of the tracking of the reference particles
+            in the plot.
+        """
+        reference_segment = deepcopy(self)
+        splits = reference_segment.split(resolution=torch.tensor(resolution))
+
+        split_lengths = [split.length[0] for split in splits]
+        ss = [0] + [sum(split_lengths[: i + 1]) for i, _ in enumerate(split_lengths)]
+
+        references = []
+        if beam is None:
+            initial = ParticleBeam.make_linspaced(
+                num_particles=num_particles, device="cpu"
+            )
+            references.append(initial)
+        else:
+            initial = ParticleBeam.make_linspaced(
+                num_particles=num_particles,
+                mu_x=beam.mu_x,
+                mu_xp=beam.mu_xp,
+                mu_y=beam.mu_y,
+                mu_yp=beam.mu_yp,
+                sigma_x=beam.sigma_x,
+                sigma_xp=beam.sigma_xp,
+                sigma_y=beam.sigma_y,
+                sigma_yp=beam.sigma_yp,
+                sigma_s=beam.sigma_s,
+                sigma_p=beam.sigma_p,
+                energy=beam.energy,
+                device="cpu",
+            )
+            references.append(initial)
+        for split in splits:
+            sample = split(references[-1])
+            references.append(sample)
+
+        for particle_index in range(num_particles):
+            xs = [
+                float(reference_beam.xs[0, particle_index].cpu())
+                for reference_beam in references
+                if reference_beam is not Beam.empty
+            ]
+            axx.plot(ss[: len(xs)], xs)
+        axx.set_xlabel("s (m)")
+        axx.set_ylabel("x (m)")
+        axx.grid()
+
+        for particle_index in range(num_particles):
+            ys = [
+                float(reference_beam.ys[0, particle_index].cpu())
+                for reference_beam in references
+                if reference_beam is not Beam.empty
+            ]
+            axy.plot(ss[: len(ys)], ys)
+        axx.set_xlabel("s (m)")
+        axy.set_ylabel("y (m)")
+        axy.grid()
+
+    def plot_overview(
+        self,
+        fig: Optional[matplotlib.figure.Figure] = None,
+        beam: Optional[Beam] = None,
+        n: int = 10,
+        resolution: float = 0.01,
+    ) -> None:
+        """
+        Plot an overview of the segment with the lattice and traced reference particles.
+
+        :param fig: Figure to plot the overview into.
+        :param beam: Entering beam from which the reference particles are sampled.
+        :param n: Number of reference particles to plot. Must not be larger than number
+            of particles passed in `beam`.
+        :param resolution: Minimum resolution of the tracking of the reference particles
+            in the plot.
+        """
+        if fig is None:
+            fig = plt.figure()
+        gs = fig.add_gridspec(3, hspace=0, height_ratios=[2, 2, 1])
+        axs = gs.subplots(sharex=True)
+
+        axs[0].set_title("Reference Particle Traces")
+        self.plot_reference_particle_traces(axs[0], axs[1], beam, n, resolution)
+
+        self.plot(axs[2], 0)
+
+        plt.tight_layout()
+
+    def plot_twiss(self, beam: Beam, ax: Optional[Any] = None) -> None:
+        """Plot twiss parameters along the segment."""
+        longitudinal_beams = [beam]
+        s_positions = [0.0]
+        for element in self.elements:
+            if element.length == 0:
+                continue
+
+            outgoing = element.track(longitudinal_beams[-1])
+
+            longitudinal_beams.append(outgoing)
+            s_positions.append(s_positions[-1] + element.length)
+
+        beta_x = [beam.beta_x for beam in longitudinal_beams]
+        beta_y = [beam.beta_y for beam in longitudinal_beams]
+
+        if ax is None:
+            fig = plt.figure()
+            ax = fig.add_subplot(111)
+
+        ax.set_title("Twiss Parameters")
+        ax.set_xlabel("s (m)")
+        ax.set_ylabel(r"$\beta$ (m)")
+
+        ax.plot(s_positions, beta_x, label=r"$\beta_x$", c="tab:red")
+        ax.plot(s_positions, beta_y, label=r"$\beta_y$", c="tab:green")
+
+        ax.legend()
+        plt.tight_layout()
+
+    @property
+    def defining_features(self) -> list[str]:
+        return super().defining_features + ["elements"]
+
+    def plot_twiss_over_lattice(self, beam: Beam, figsize=(8, 4)) -> None:
+        """Plot twiss parameters in a plot over a plot of the lattice."""
+        fig = plt.figure(figsize=figsize)
+        gs = fig.add_gridspec(2, hspace=0, height_ratios=[3, 1])
+        axs = gs.subplots(sharex=True)
+
+        self.plot_twiss(beam, ax=axs[0])
+        self.plot(axs[1], 0)
+
+        plt.tight_layout()
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}(elements={repr(self.elements)}, "
+            + f"name={repr(self.name)})"
+        )
