+++ cheetah/particles/__init__.py
@@ -1,0 +1,3 @@
+from .beam import Beam  # noqa: F401
+from .parameter_beam import ParameterBeam  # noqa: F401
+from .particle_beam import ParticleBeam  # noqa: F401
