+++ docs/accelerator.rst
@@ -1,8 +1,68 @@
-.. Documents accelerator.py
+.. Documents accelerator subpackage
 
 Accelerator
 ===========
 
-.. automodule:: accelerator
+.. automodule:: accelerator.aperture
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.bpm
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.cavity
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.custom_transfer_map
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.dipole
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.drift
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.element
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.horizontal_corrector
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.marker
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.quadrupole
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.rbend
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.screen
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.segment
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.solenoid
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.undulator
+    :members:
+    :undoc-members:
+
+.. automodule:: accelerator.vertical_corrector
     :members:
     :undoc-members:
