+++ docs/particles.rst
@@ -1,8 +1,16 @@
-.. Documents particles.py
+.. Documents particles subpackage
 
 Particles
 =========
 
-.. automodule:: particles
+.. automodule:: particles.beam
+    :members:
+    :undoc-members:
+
+.. automodule:: particles.parameter_beam
+    :members:
+    :undoc-members:
+
+.. automodule:: particles.particle_beam
     :members:
     :undoc-members:
